Add new AdditionalConfig property to routers and interfaces
